+++ lib/flatplan/ports.rb
@@ -1,5 +1,6 @@
 require_relative 'ports/context'
 require_relative 'ports/file_store'
+require_relative 'ports/metadata_store'
 
 module Flatplan